+++ lib/main.dart
@@ -1,19 +1,16 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:bot_toast/bot_toast.dart';
-
 import 'package:e_commerce/src/config/routes/auto_router.gr.dart';
 import 'package:e_commerce/src/config/themes/themes.dart';
-
+import 'package:e_commerce/src/models/handle.dart';
 import 'package:e_commerce/src/models/users_model.dart';
 import 'package:e_commerce/src/modules/account/logic/account_bloc.dart';
 import 'package:e_commerce/src/modules/favorites/logic/favorites_bloc.dart';
-import 'package:e_commerce/src/modules/home/logic/product_bloc.dart';
-import 'package:e_commerce/src/modules/product_by_category/logic/product_by_category_bloc.dart';
+import 'package:e_commerce/src/modules/product/logic/product_bloc.dart';
 import 'package:e_commerce/src/modules/shop/logic/categories_bloc.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:get_it/get_it.dart';
 
@@ -42,9 +39,11 @@ class MyApp extends StatelessWidget {
       providers: [
         BlocProvider(
             create: (_) => AccountBloc(AccountState(data: XUser.empty()))),
-        BlocProvider(create: (_) => ProductBloc()),
+        BlocProvider(
+            create: (_) => ProductBloc(ProductState(
+                searchList: XHandle.completed([]),
+                items: XHandle.completed([])))),
         BlocProvider(create: (_) => CategoriesBloc()),
-        BlocProvider(create: (_) => ProductByCategoryBloc()),
         BlocProvider(create: (_) => FavoriteBloc()),
       ],
       child: MaterialApp.router(
@@ -64,7 +63,9 @@ class MyApp extends StatelessWidget {
           Locale('en', ''),
         ],
         onGenerateTitle: (BuildContext context) => S.of(context).appTitle,
-        routeInformationParser: appRouter.defaultRouteParser(),
+        routeInformationParser: appRouter.defaultRouteParser(
+          includePrefixMatches: true,
+        ),
         routerDelegate: AutoRouterDelegate(appRouter),
       ),
     );
