+++ lib/src/config/routes/coordinator.dart
@@ -1,6 +1,7 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:e_commerce/src/config/routes/auto_router.gr.dart';
 import 'package:e_commerce/src/models/products_model.dart';
+import 'package:e_commerce/src/modules/product_details/pages/product_details_page.dart';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
 
@@ -16,7 +17,11 @@ class XCoordinator {
   static showDashboard() {
     rootRouter.replace(const DashboardWrapperRoute());
   }
-   static showDetailProduct(BuildContext context,{required XProduct data}) {
-    context.router.push( ProductDetailsRoute(data:data ));
+
+  static showDetailProduct(BuildContext context, {required XProduct data}) {
+    context.router.pushWidget(ProductDetailsPage(
+      id: data.id,
+      data: data,
+    ));
   }
 }
