+++ lib/src/config/routes/page_routers.dart
@@ -2,4 +2,5 @@ class XRoutes {
   static const String dashboard = "/dashboard";
   static const String loading = "";
   static const String sign = "/sign";
+  static const String detailProduct = '/detail/:id';
 }
