+++ lib/src/modules/detail_category/widgets/header_detail_category_delegate.dart
@@ -1,13 +1,18 @@
 import 'package:e_commerce/src/config/themes/my_colors.dart';
 import 'package:e_commerce/src/modules/dashboard/router/dashboard_router.dart';
-import 'package:e_commerce/src/modules/favorites/widgets/filter_bar.dart';
+import 'package:e_commerce/src/modules/detail_category/widgets/filter_bar_detail_category.dart';
 import 'package:e_commerce/src/modules/shop/logic/categories_bloc.dart';
+
 import 'package:e_commerce/src/widgets/chip/tag_chip.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-class HeaderFavorite extends SliverPersistentHeaderDelegate {
-  const HeaderFavorite();
+class HeaderDetailCategory extends SliverPersistentHeaderDelegate {
+  final String nameCategory;
+
+  const HeaderDetailCategory({
+    required this.nameCategory,
+  });
 
   @override
   Widget build(
@@ -36,7 +41,7 @@ class HeaderFavorite extends SliverPersistentHeaderDelegate {
                   icon: const Icon(Icons.search),
                   color: MyColors.colorBlack,
                   onPressed: () =>
-                      DashboardCoordinator.showSearchProductByFavorite(context),
+                      DashboardCoordinator.showSearchProductByCategory(context),
                 )
               ],
               elevation: 3,
@@ -55,7 +60,7 @@ class HeaderFavorite extends SliverPersistentHeaderDelegate {
               child: Opacity(
                 opacity: shrinkOffset > 1 && shrinkOffset < 100 ? progress : 1,
                 child: Text(
-                  'Favorites',
+                  nameCategory,
                   textAlign: TextAlign.center,
                   style: TextStyle.lerp(
                     Theme.of(context).textTheme.headlineLarge,
@@ -96,7 +101,6 @@ class HeaderFavorite extends SliverPersistentHeaderDelegate {
                             return TagChip(
                               label: items[index].name ?? 'N/A',
                               idCategory: items[index].id,
-                              onPressed: () {},
                             );
                           },
                         ),
@@ -104,7 +108,7 @@ class HeaderFavorite extends SliverPersistentHeaderDelegate {
                     }),
                     const Padding(
                       padding: EdgeInsets.fromLTRB(14, 5, 14, 0),
-                      child: FilerBarFavorites(),
+                      child: FilerBarDetailCategory(),
                     ),
                   ],
                 ),
