+++ lib/src/modules/favorites/logic/favorites_bloc.dart
@@ -1,40 +1,39 @@
 import 'package:e_commerce/src/config/routes/coordinator.dart';
-import 'package:e_commerce/src/config/themes/my_colors.dart';
 import 'package:e_commerce/src/models/handle.dart';
 import 'package:e_commerce/src/models/products_model.dart';
-import 'package:e_commerce/src/modules/product_by_category/logic/product_by_category_bloc.dart';
-import 'package:e_commerce/src/repositories/domain.dart';
+import 'package:e_commerce/src/modules/product/logic/product_bloc.dart';
 import 'package:e_commerce/src/repositories/firestore/services/auth_service.dart';
+import 'package:e_commerce/src/utils/enum/size_type.dart';
+import 'package:e_commerce/src/utils/enum/sort_by.dart';
+import 'package:e_commerce/src/utils/enum/view_type.dart';
 import 'package:e_commerce/src/widgets/snackbar/snackbar.dart';
-import 'package:equatable/equatable.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 
 part 'favorites_state.dart';
 
-class FavoriteBloc extends Cubit<FavoriteState> {
+class FavoriteBloc extends ProductBloc<FavoriteState> {
   FavoriteBloc()
       : super(FavoriteState(
-            favoriteList: XHandle.completed([]),
+            listFavorite: XHandle.completed([]),
+            items: XHandle.completed([]),
             searchList: XHandle.completed([]))) {
-    getProduct();
+    getFavorite();
   }
-  final Domain domain = Domain();
 
-  Future<void> getProduct() async {
+  Future<void> getFavorite() async {
     User? currentUser = AuthService().currentUser;
     final value = await domain.favorite.getProductToFavorite();
     if (value.isSuccess) {
-      List<XProduct> items = [...(state.favoriteList.data ?? [])];
+      List<XProduct> items = [...(state.listFavorite.data ?? [])];
       items = (value.data ?? [])
           .where((e) => e.idUser == currentUser?.uid)
           .toList();
-      emit(state.copyWithItem(favoriteList: XHandle.completed(items)));
+      emit(state.copyWithItem(listFavorite: XHandle.completed(items)));
     } else {}
   }
 
-  Future<void> addProduct(BuildContext context,
+  Future<void> addProductToFavorite(BuildContext context,
       {required XProduct product, required SizeType sizeType}) async {
     XProduct xProduct = XProduct(
       color: product.color,
@@ -56,11 +55,11 @@ class FavoriteBloc extends Cubit<FavoriteState> {
     final value = await domain.favorite.addProductToFavorite(xProduct);
     if (value.isSuccess) {
       final List<XProduct> items = [
-        ...(state.favoriteList.data ?? []),
+        ...(state.listFavorite.data ?? []),
         xProduct
       ];
 
-      emit(state.copyWithItem(favoriteList: XHandle.completed(items)));
+      emit(state.copyWithItem(listFavorite: XHandle.completed(items)));
       XSnackBar.show(msg: 'Favorite success');
       XCoordinator.pop(context);
     } else {
@@ -68,47 +67,25 @@ class FavoriteBloc extends Cubit<FavoriteState> {
     }
   }
 
-  Future<void> removeProduct(XProduct product) async {
+  Future<void> removeProductToFavorite(XProduct product) async {
     final value = await domain.favorite.deleteProductToFavorite(product);
     if (value.isSuccess) {
-      final List<XProduct> items = [...(state.favoriteList.data ?? [])];
+      final List<XProduct> items = [...(state.listFavorite.data ?? [])];
       items.remove(product);
 
-      emit(state.copyWithItem(favoriteList: XHandle.completed(items)));
+      emit(state.copyWithItem(listFavorite: XHandle.completed(items)));
       XSnackBar.show(msg: 'Remove Product success');
     } else {
       XSnackBar.show(msg: 'Remove Product failure');
     }
   }
-
-  void sortBy(int index) {
-    emit(state.copyWithItem(
-      sortBy: SortBy.values[index],
-    ));
-  }
-
-  void changeViewType() {
-    var viewType = state.viewType == ViewType.listView
-        ? ViewType.gridView
-        : ViewType.listView;
-    emit(state.copyWithItem(viewType: viewType));
-  }
-
-  void onSelectSize(int index) {
-    emit(state.copyWithItem(sizeType: SizeType.values[index]));
-  }
-
-  void initSizeType() async {
-    await Future.delayed(const Duration(seconds: 2));
-    emit(state.copyWithItem(sizeType: SizeType.xs));
-  }
-
+    @override
   Future<void> searchProduct(String query) async {
     late List<XProduct> items;
     if (query.isEmpty || query == '') {
       items = [];
     } else {
-      items = (state.favoriteList.data ?? []).where((e) {
+      items = (state.listFavorite.data ?? []).where((e) {
         final titleLower = e.name.toLowerCase();
         final searchLower = query.toLowerCase();
