+++ lib/src/modules/favorites/pages/search_favorite_page.dart
@@ -3,14 +3,14 @@ import 'package:e_commerce/src/config/themes/my_colors.dart';
 import 'package:e_commerce/src/models/handle.dart';
 import 'package:e_commerce/src/models/result.dart';
 import 'package:e_commerce/src/modules/favorites/logic/favorites_bloc.dart';
-import 'package:e_commerce/src/modules/product_by_category/widgets/search_text_filed.dart';
+import 'package:e_commerce/src/widgets/text_field/search_text_filed.dart';
 import 'package:e_commerce/src/widgets/state/state_error_widget.dart';
 import 'package:e_commerce/src/widgets/state/state_loading_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-class SearchProductsByFavoritePage extends StatelessWidget {
-  const SearchProductsByFavoritePage({Key? key}) : super(key: key);
+class SearchFavoritePage extends StatelessWidget {
+  const SearchFavoritePage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
