+++ lib/src/modules/favorites/widgets/product_card_horizontal.dart
@@ -17,9 +17,6 @@ class XProductCardFavoriteHorizontal extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    var numberStarActive = ((data.star).toInt() ~/ 5);
-    var numberStarNoActive = 5 - numberStarActive;
-
     return data.soldOut
         ? _soldOutWidget()
         : GestureDetector(
@@ -29,188 +26,7 @@ class XProductCardFavoriteHorizontal extends StatelessWidget {
               width: 343,
               child: Stack(
                 children: [
-                  Container(
-                    height: 94,
-                    decoration: const BoxDecoration(
-                        color: MyColors.colorWhite,
-                        boxShadow: [
-                          BoxShadow(
-                              blurRadius: 25,
-                              offset: Offset(0, 1),
-                              color: MyColors.colorWhite,
-                              spreadRadius: 1)
-                        ],
-                        borderRadius: BorderRadius.all(Radius.circular(8))),
-                    child: Row(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        SizedBox(
-                          height: 104,
-                          width: 104,
-                          child: ClipRRect(
-                            borderRadius: const BorderRadius.only(
-                                topLeft: Radius.circular(8),
-                                bottomLeft: Radius.circular(8)),
-                            child: Image(
-                              fit: BoxFit.fill,
-                              image: NetworkImage(data.image ?? "N/A"),
-                            ),
-                          ),
-                        ),
-                        const SizedBox(
-                          width: 11,
-                        ),
-                        Column(
-                          mainAxisAlignment: MainAxisAlignment.center,
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            Text(
-                              data.name,
-                              style: const TextStyle(
-                                  fontSize: 11,
-                                  color: MyColors.colorGray,
-                                  fontWeight: FontWeight.normal),
-                            ),
-                            Text(
-                              data.type,
-                              style: const TextStyle(
-                                  fontSize: 16,
-                                  color: MyColors.colorBlack,
-                                  height: 1,
-                                  fontWeight: FontWeight.w600),
-                            ),
-                            Row(
-                              children: [
-                                RichText(
-                                  text: TextSpan(
-                                    text: 'Color: ',
-                                    style: const TextStyle(
-                                        fontSize: 11,
-                                        height: 1,
-                                        fontWeight: FontWeight.normal,
-                                        color: MyColors.colorGray),
-                                    children: <TextSpan>[
-                                      TextSpan(
-                                          text: '${data.color}',
-                                          style: const TextStyle(
-                                              color: MyColors.colorBlack)),
-                                    ],
-                                  ),
-                                ),
-                                const SizedBox(
-                                  width: 60,
-                                ),
-                                RichText(
-                                  text: TextSpan(
-                                    text: 'Size: ',
-                                    style: const TextStyle(
-                                        fontSize: 11,
-                                        height: 1,
-                                        fontWeight: FontWeight.normal,
-                                        color: MyColors.colorGray),
-                                    children: <TextSpan>[
-                                      TextSpan(
-                                          text: '${data.size}',
-                                          style: const TextStyle(
-                                              color: MyColors.colorBlack)),
-                                    ],
-                                  ),
-                                )
-                              ],
-                            ),
-                            const SizedBox(height: 5),
-                            Row(
-                              children: [
-                                SizedBox(
-                                  width: 60,
-                                  child: data.discount == 0.0
-                                      ? Text(
-                                          "${XUtils.formatPrice(data.originalPrice)}\$ ",
-                                          style: const TextStyle(
-                                              fontSize: 14,
-                                              height: 1.42,
-                                              color: MyColors.colorBlack,
-                                              fontWeight: FontWeight.w600),
-                                        )
-                                      : RichText(
-                                          text: TextSpan(
-                                            text:
-                                                "${XUtils.formatPrice(data.originalPrice)}\$ ",
-                                            style: const TextStyle(
-                                                fontSize: 14,
-                                                height: 1.42,
-                                                decoration:
-                                                    TextDecoration.lineThrough,
-                                                color: MyColors.colorGray,
-                                                fontWeight: FontWeight.w600),
-                                            children: <TextSpan>[
-                                              TextSpan(
-                                                  text:
-                                                      "${XUtils.formatPrice(data.currentPrice ?? -1)}\$",
-                                                  style: const TextStyle(
-                                                    color:
-                                                        MyColors.colorSaleHot,
-                                                    decoration:
-                                                        TextDecoration.none,
-                                                  )),
-                                            ],
-                                          ),
-                                        ),
-                                ),
-                                const SizedBox(
-                                  width: 30,
-                                ),
-                                Row(
-                                  children: [
-                                    SizedBox(
-                                      height: 20,
-                                      child: ListView.builder(
-                                          scrollDirection: Axis.horizontal,
-                                          shrinkWrap: true,
-                                          itemBuilder: (context, index) {
-                                            return SizedBox(
-                                              width: 13,
-                                              height: 12,
-                                              child: Image.asset(
-                                                  MyIcons.activeStarIcon,
-                                                  fit: BoxFit.scaleDown),
-                                            );
-                                          },
-                                          itemCount: numberStarActive),
-                                    ),
-                                    SizedBox(
-                                      height: 20,
-                                      child: ListView.builder(
-                                          scrollDirection: Axis.horizontal,
-                                          shrinkWrap: true,
-                                          itemBuilder: (context, index) {
-                                            return SizedBox(
-                                              width: 13,
-                                              height: 12,
-                                              child: Image.asset(
-                                                  MyIcons.starIcon,
-                                                  fit: BoxFit.scaleDown),
-                                            );
-                                          },
-                                          itemCount: numberStarNoActive),
-                                    ),
-                                    const SizedBox(width: 3),
-                                    Text(
-                                      "(${data.star})",
-                                      style: const TextStyle(
-                                          fontSize: 10,
-                                          color: MyColors.colorGray,
-                                          fontWeight: FontWeight.normal),
-                                    )
-                                  ],
-                                ),
-                              ],
-                            ),
-                          ],
-                        ),
-                      ],
-                    ),
-                  ),
+                  _belowCard(),
                   SizedBox(
                     height: 104,
                     child: Column(
@@ -236,7 +52,7 @@ class XProductCardFavoriteHorizontal extends StatelessWidget {
                                           color: MyColors.colorGray),
                                       onTap: () => context
                                           .read<FavoriteBloc>()
-                                          .removeProduct(data),
+                                          .removeProductToFavorite(data),
                                     )),
                               )
                             ],
@@ -256,191 +72,192 @@ class XProductCardFavoriteHorizontal extends StatelessWidget {
           );
   }
 
-  Widget _soldOutWidget() {
+  Widget _belowCard() {
     var numberStarActive = ((data.star).toInt() ~/ 5);
     var numberStarNoActive = 5 - numberStarActive;
+    return Container(
+      height: 94,
+      decoration: const BoxDecoration(
+          color: MyColors.colorWhite,
+          boxShadow: [
+            BoxShadow(
+                blurRadius: 25,
+                offset: Offset(0, 1),
+                color: MyColors.colorWhite,
+                spreadRadius: 1)
+          ],
+          borderRadius: BorderRadius.all(Radius.circular(8))),
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          SizedBox(
+            height: 104,
+            width: 104,
+            child: ClipRRect(
+              borderRadius: const BorderRadius.only(
+                  topLeft: Radius.circular(8), bottomLeft: Radius.circular(8)),
+              child: Image(
+                fit: BoxFit.fill,
+                image: NetworkImage(data.image ?? "N/A"),
+              ),
+            ),
+          ),
+          const SizedBox(
+            width: 11,
+          ),
+          Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text(
+                data.name,
+                style: const TextStyle(
+                    fontSize: 11,
+                    color: MyColors.colorGray,
+                    fontWeight: FontWeight.normal),
+              ),
+              Text(
+                data.type,
+                style: const TextStyle(
+                    fontSize: 16,
+                    color: MyColors.colorBlack,
+                    height: 1,
+                    fontWeight: FontWeight.w600),
+              ),
+              Row(
+                children: [
+                  RichText(
+                    text: TextSpan(
+                      text: 'Color: ',
+                      style: const TextStyle(
+                          fontSize: 11,
+                          height: 1,
+                          fontWeight: FontWeight.normal,
+                          color: MyColors.colorGray),
+                      children: <TextSpan>[
+                        TextSpan(
+                            text: '${data.color}',
+                            style: const TextStyle(color: MyColors.colorBlack)),
+                      ],
+                    ),
+                  ),
+                  const SizedBox(
+                    width: 60,
+                  ),
+                  RichText(
+                    text: TextSpan(
+                      text: 'Size: ',
+                      style: const TextStyle(
+                          fontSize: 11,
+                          height: 1,
+                          fontWeight: FontWeight.normal,
+                          color: MyColors.colorGray),
+                      children: <TextSpan>[
+                        TextSpan(
+                            text: '${data.size}',
+                            style: const TextStyle(color: MyColors.colorBlack)),
+                      ],
+                    ),
+                  )
+                ],
+              ),
+              const SizedBox(height: 5),
+              Row(
+                children: [
+                  SizedBox(
+                    width: 60,
+                    child: data.discount == 0.0
+                        ? Text(
+                            "${XUtils.formatPrice(data.originalPrice)}\$ ",
+                            style: const TextStyle(
+                                fontSize: 14,
+                                height: 1.42,
+                                color: MyColors.colorBlack,
+                                fontWeight: FontWeight.w600),
+                          )
+                        : RichText(
+                            text: TextSpan(
+                              text:
+                                  "${XUtils.formatPrice(data.originalPrice)}\$ ",
+                              style: const TextStyle(
+                                  fontSize: 14,
+                                  height: 1.42,
+                                  decoration: TextDecoration.lineThrough,
+                                  color: MyColors.colorGray,
+                                  fontWeight: FontWeight.w600),
+                              children: <TextSpan>[
+                                TextSpan(
+                                    text:
+                                        "${XUtils.formatPrice(data.currentPrice ?? -1)}\$",
+                                    style: const TextStyle(
+                                      color: MyColors.colorSaleHot,
+                                      decoration: TextDecoration.none,
+                                    )),
+                              ],
+                            ),
+                          ),
+                  ),
+                  const SizedBox(
+                    width: 30,
+                  ),
+                  Row(
+                    children: [
+                      SizedBox(
+                        height: 20,
+                        child: ListView.builder(
+                            scrollDirection: Axis.horizontal,
+                            shrinkWrap: true,
+                            itemBuilder: (context, index) {
+                              return SizedBox(
+                                width: 13,
+                                height: 12,
+                                child: Image.asset(MyIcons.activeStarIcon,
+                                    fit: BoxFit.scaleDown),
+                              );
+                            },
+                            itemCount: numberStarActive),
+                      ),
+                      SizedBox(
+                        height: 20,
+                        child: ListView.builder(
+                            scrollDirection: Axis.horizontal,
+                            shrinkWrap: true,
+                            itemBuilder: (context, index) {
+                              return SizedBox(
+                                width: 13,
+                                height: 12,
+                                child: Image.asset(MyIcons.starIcon,
+                                    fit: BoxFit.scaleDown),
+                              );
+                            },
+                            itemCount: numberStarNoActive),
+                      ),
+                      const SizedBox(width: 3),
+                      Text(
+                        "(${data.star})",
+                        style: const TextStyle(
+                            fontSize: 10,
+                            color: MyColors.colorGray,
+                            fontWeight: FontWeight.normal),
+                      )
+                    ],
+                  ),
+                ],
+              ),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _soldOutWidget() {
     return SizedBox(
       height: 114,
       width: 343,
       child: Stack(
         children: [
-          Container(
-            height: 94,
-            decoration: const BoxDecoration(
-                color: MyColors.colorWhite,
-                boxShadow: [
-                  BoxShadow(
-                      blurRadius: 25,
-                      offset: Offset(0, 1),
-                      color: MyColors.colorWhite,
-                      spreadRadius: 1)
-                ],
-                borderRadius: BorderRadius.all(Radius.circular(8))),
-            child: Row(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                SizedBox(
-                  height: 104,
-                  width: 104,
-                  child: ClipRRect(
-                    borderRadius: const BorderRadius.only(
-                        topLeft: Radius.circular(8),
-                        bottomLeft: Radius.circular(8)),
-                    child: Image(
-                      fit: BoxFit.fill,
-                      image: NetworkImage(data.image ?? "N/A"),
-                    ),
-                  ),
-                ),
-                const SizedBox(
-                  width: 11,
-                ),
-                Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      data.name,
-                      style: const TextStyle(
-                          fontSize: 11,
-                          color: MyColors.colorGray,
-                          fontWeight: FontWeight.normal),
-                    ),
-                    Text(
-                      data.type,
-                      style: const TextStyle(
-                          fontSize: 16,
-                          color: MyColors.colorBlack,
-                          height: 1,
-                          fontWeight: FontWeight.w600),
-                    ),
-                    Row(
-                      children: [
-                        RichText(
-                          text: TextSpan(
-                            text: 'Color: ',
-                            style: const TextStyle(
-                                fontSize: 11,
-                                height: 1,
-                                fontWeight: FontWeight.normal,
-                                color: MyColors.colorGray),
-                            children: <TextSpan>[
-                              TextSpan(
-                                  text: '${data.color}',
-                                  style: const TextStyle(
-                                      color: MyColors.colorBlack)),
-                            ],
-                          ),
-                        ),
-                        const SizedBox(
-                          width: 60,
-                        ),
-                        RichText(
-                          text: TextSpan(
-                            text: 'Size: ',
-                            style: const TextStyle(
-                                fontSize: 11,
-                                height: 1,
-                                fontWeight: FontWeight.normal,
-                                color: MyColors.colorGray),
-                            children: <TextSpan>[
-                              TextSpan(
-                                  text: '${data.size}',
-                                  style: const TextStyle(
-                                      color: MyColors.colorBlack)),
-                            ],
-                          ),
-                        )
-                      ],
-                    ),
-                    const SizedBox(height: 5),
-                    Row(
-                      children: [
-                        SizedBox(
-                          width: 60,
-                          child: data.discount == 0.0
-                              ? Text(
-                                  "${XUtils.formatPrice(data.originalPrice)}\$ ",
-                                  style: const TextStyle(
-                                      fontSize: 14,
-                                      height: 1.42,
-                                      color: MyColors.colorBlack,
-                                      fontWeight: FontWeight.w600),
-                                )
-                              : RichText(
-                                  text: TextSpan(
-                                    text:
-                                        "${XUtils.formatPrice(data.originalPrice)}\$ ",
-                                    style: const TextStyle(
-                                        fontSize: 14,
-                                        height: 1.42,
-                                        decoration: TextDecoration.lineThrough,
-                                        color: MyColors.colorGray,
-                                        fontWeight: FontWeight.w600),
-                                    children: <TextSpan>[
-                                      TextSpan(
-                                          text:
-                                              "${XUtils.formatPrice(data.currentPrice ?? -1)}\$",
-                                          style: const TextStyle(
-                                            color: MyColors.colorSaleHot,
-                                            decoration: TextDecoration.none,
-                                          )),
-                                    ],
-                                  ),
-                                ),
-                        ),
-                        const SizedBox(
-                          width: 30,
-                        ),
-                        Row(
-                          children: [
-                            SizedBox(
-                              height: 20,
-                              child: ListView.builder(
-                                  scrollDirection: Axis.horizontal,
-                                  shrinkWrap: true,
-                                  itemBuilder: (context, index) {
-                                    return SizedBox(
-                                      width: 13,
-                                      height: 12,
-                                      child: Image.asset(MyIcons.activeStarIcon,
-                                          fit: BoxFit.scaleDown),
-                                    );
-                                  },
-                                  itemCount: numberStarActive),
-                            ),
-                            SizedBox(
-                              height: 20,
-                              child: ListView.builder(
-                                  scrollDirection: Axis.horizontal,
-                                  shrinkWrap: true,
-                                  itemBuilder: (context, index) {
-                                    return SizedBox(
-                                      width: 13,
-                                      height: 12,
-                                      child: Image.asset(MyIcons.starIcon,
-                                          fit: BoxFit.scaleDown),
-                                    );
-                                  },
-                                  itemCount: numberStarNoActive),
-                            ),
-                            const SizedBox(width: 3),
-                            Text(
-                              "(${data.star})",
-                              style: const TextStyle(
-                                  fontSize: 10,
-                                  color: MyColors.colorGray,
-                                  fontWeight: FontWeight.normal),
-                            )
-                          ],
-                        ),
-                      ],
-                    ),
-                  ],
-                ),
-              ],
-            ),
-          ),
+          _belowCard(),
           Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
@@ -470,7 +287,7 @@ class XProductCardFavoriteHorizontal extends StatelessWidget {
                                       color: MyColors.colorGray),
                                   onTap: () => context
                                       .read<FavoriteBloc>()
-                                      .removeProduct(data),
+                                      .removeProductToFavorite(data),
                                 )),
                           )
                         ],
