+++ lib/src/modules/favorites/widgets/product_card_vertical.dart
@@ -17,8 +17,6 @@ class XProductCardFavoriteVertical extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    var numberStarActive = ((data.star).toInt() ~/ 5);
-    var numberStarNoActive = 5 - numberStarActive;
     return data.soldOut
         ? _soldOutWidget()
         : GestureDetector(
@@ -28,150 +26,7 @@ class XProductCardFavoriteVertical extends StatelessWidget {
               width: 164,
               child: Stack(
                 children: [
-                  Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      SizedBox(
-                        height: 184,
-                        width: 162,
-                        child: ClipRRect(
-                          borderRadius: BorderRadius.circular(8),
-                          child: Image(
-                            fit: BoxFit.fill,
-                            image: NetworkImage(data.image ?? "N/A"),
-                          ),
-                        ),
-                      ),
-                      Row(
-                        children: [
-                          SizedBox(
-                            height: 20,
-                            child: ListView.builder(
-                                scrollDirection: Axis.horizontal,
-                                shrinkWrap: true,
-                                itemBuilder: (context, index) {
-                                  return SizedBox(
-                                    width: 13,
-                                    height: 12,
-                                    child: Image.asset(MyIcons.activeStarIcon,
-                                        fit: BoxFit.scaleDown),
-                                  );
-                                },
-                                itemCount: numberStarActive),
-                          ),
-                          SizedBox(
-                            height: 20,
-                            child: ListView.builder(
-                                scrollDirection: Axis.horizontal,
-                                shrinkWrap: true,
-                                itemBuilder: (context, index) {
-                                  return SizedBox(
-                                    width: 13,
-                                    height: 12,
-                                    child: Image.asset(MyIcons.starIcon,
-                                        fit: BoxFit.scaleDown),
-                                  );
-                                },
-                                itemCount: numberStarNoActive),
-                          ),
-                          const SizedBox(width: 3),
-                          Text(
-                            "${data.star}",
-                            style: const TextStyle(
-                                fontSize: 10,
-                                color: MyColors.colorGray,
-                                fontWeight: FontWeight.normal),
-                          )
-                        ],
-                      ),
-                      Text(
-                        data.name,
-                        style: const TextStyle(
-                            fontSize: 11,
-                            color: MyColors.colorGray,
-                            fontWeight: FontWeight.normal),
-                      ),
-                      Text(
-                        data.type,
-                        maxLines: 1,
-                        style: const TextStyle(
-                            fontSize: 16,
-                            overflow: TextOverflow.clip,
-                            color: MyColors.colorBlack,
-                            height: 1,
-                            fontWeight: FontWeight.w600),
-                      ),
-                      Row(
-                        children: [
-                          RichText(
-                            text: TextSpan(
-                              text: 'Color: ',
-                              style: const TextStyle(
-                                  fontSize: 11,
-                                  height: 1,
-                                  fontWeight: FontWeight.normal,
-                                  color: MyColors.colorGray),
-                              children: <TextSpan>[
-                                TextSpan(
-                                    text: '${data.color}',
-                                    style: const TextStyle(
-                                        color: MyColors.colorBlack)),
-                              ],
-                            ),
-                          ),
-                          const SizedBox(
-                            width: 30,
-                          ),
-                          RichText(
-                            text: TextSpan(
-                              text: 'Size: ',
-                              style: const TextStyle(
-                                  fontSize: 11,
-                                  height: 1,
-                                  fontWeight: FontWeight.normal,
-                                  color: MyColors.colorGray),
-                              children: <TextSpan>[
-                                TextSpan(
-                                    text: '${data.size}',
-                                    style: const TextStyle(
-                                        color: MyColors.colorBlack)),
-                              ],
-                            ),
-                          )
-                        ],
-                      ),
-                      data.discount == 0.0
-                          ? Text(
-                              "${XUtils.formatPrice(data.originalPrice)}\$ ",
-                              style: const TextStyle(
-                                  fontSize: 14,
-                                  height: 1.42,
-                                  color: MyColors.colorBlack,
-                                  fontWeight: FontWeight.w600),
-                            )
-                          : RichText(
-                              text: TextSpan(
-                                text:
-                                    "${XUtils.formatPrice(data.originalPrice)}\$ ",
-                                style: const TextStyle(
-                                    fontSize: 14,
-                                    height: 1.42,
-                                    decoration: TextDecoration.lineThrough,
-                                    color: MyColors.colorGray,
-                                    fontWeight: FontWeight.w600),
-                                children: <TextSpan>[
-                                  TextSpan(
-                                      text:
-                                          "${XUtils.formatPrice(data.currentPrice ?? -1)}\$",
-                                      style: const TextStyle(
-                                        color: MyColors.colorSaleHot,
-                                        decoration: TextDecoration.none,
-                                      )),
-                                ],
-                              ),
-                            ),
-                    ],
-                  ),
+                  _belowCard(),
                   SizedBox(
                     height: 200,
                     child: Column(
@@ -197,7 +52,7 @@ class XProductCardFavoriteVertical extends StatelessWidget {
                                           color: MyColors.colorGray),
                                       onTap: () => context
                                           .read<FavoriteBloc>()
-                                          .removeProduct(data),
+                                          .removeProductToFavorite(data),
                                     )),
                               )
                             ],
@@ -217,155 +72,159 @@ class XProductCardFavoriteVertical extends StatelessWidget {
           );
   }
 
-  Widget _soldOutWidget() {
+  Widget _belowCard() {
     var numberStarActive = ((data.star).toInt() ~/ 5);
     var numberStarNoActive = 5 - numberStarActive;
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        SizedBox(
+          height: 184,
+          width: 162,
+          child: ClipRRect(
+            borderRadius: BorderRadius.circular(8),
+            child: Image(
+              fit: BoxFit.fill,
+              image: NetworkImage(data.image ?? "N/A"),
+            ),
+          ),
+        ),
+        Row(
+          children: [
+            SizedBox(
+              height: 20,
+              child: ListView.builder(
+                  scrollDirection: Axis.horizontal,
+                  shrinkWrap: true,
+                  itemBuilder: (context, index) {
+                    return SizedBox(
+                      width: 13,
+                      height: 12,
+                      child: Image.asset(MyIcons.activeStarIcon,
+                          fit: BoxFit.scaleDown),
+                    );
+                  },
+                  itemCount: numberStarActive),
+            ),
+            SizedBox(
+              height: 20,
+              child: ListView.builder(
+                  scrollDirection: Axis.horizontal,
+                  shrinkWrap: true,
+                  itemBuilder: (context, index) {
+                    return SizedBox(
+                      width: 13,
+                      height: 12,
+                      child:
+                          Image.asset(MyIcons.starIcon, fit: BoxFit.scaleDown),
+                    );
+                  },
+                  itemCount: numberStarNoActive),
+            ),
+            const SizedBox(width: 3),
+            Text(
+              "${data.star}",
+              style: const TextStyle(
+                  fontSize: 10,
+                  color: MyColors.colorGray,
+                  fontWeight: FontWeight.normal),
+            )
+          ],
+        ),
+        Text(
+          data.name,
+          style: const TextStyle(
+              fontSize: 11,
+              color: MyColors.colorGray,
+              fontWeight: FontWeight.normal),
+        ),
+        Text(
+          data.type,
+          maxLines: 1,
+          style: const TextStyle(
+              fontSize: 16,
+              overflow: TextOverflow.clip,
+              color: MyColors.colorBlack,
+              height: 1,
+              fontWeight: FontWeight.w600),
+        ),
+        Row(
+          children: [
+            RichText(
+              text: TextSpan(
+                text: 'Color: ',
+                style: const TextStyle(
+                    fontSize: 11,
+                    height: 1,
+                    fontWeight: FontWeight.normal,
+                    color: MyColors.colorGray),
+                children: <TextSpan>[
+                  TextSpan(
+                      text: '${data.color}',
+                      style: const TextStyle(color: MyColors.colorBlack)),
+                ],
+              ),
+            ),
+            const SizedBox(
+              width: 30,
+            ),
+            RichText(
+              text: TextSpan(
+                text: 'Size: ',
+                style: const TextStyle(
+                    fontSize: 11,
+                    height: 1,
+                    fontWeight: FontWeight.normal,
+                    color: MyColors.colorGray),
+                children: <TextSpan>[
+                  TextSpan(
+                      text: '${data.size}',
+                      style: const TextStyle(color: MyColors.colorBlack)),
+                ],
+              ),
+            )
+          ],
+        ),
+        data.discount == 0.0
+            ? Text(
+                "${XUtils.formatPrice(data.originalPrice)}\$ ",
+                style: const TextStyle(
+                    fontSize: 14,
+                    height: 1.42,
+                    color: MyColors.colorBlack,
+                    fontWeight: FontWeight.w600),
+              )
+            : RichText(
+                text: TextSpan(
+                  text: "${XUtils.formatPrice(data.originalPrice)}\$ ",
+                  style: const TextStyle(
+                      fontSize: 14,
+                      height: 1.42,
+                      decoration: TextDecoration.lineThrough,
+                      color: MyColors.colorGray,
+                      fontWeight: FontWeight.w600),
+                  children: <TextSpan>[
+                    TextSpan(
+                        text:
+                            "${XUtils.formatPrice(data.currentPrice ?? -1)}\$",
+                        style: const TextStyle(
+                          color: MyColors.colorSaleHot,
+                          decoration: TextDecoration.none,
+                        )),
+                  ],
+                ),
+              ),
+      ],
+    );
+  }
+
+  Widget _soldOutWidget() {
     return SizedBox(
       height: 280,
       width: 164,
       child: Stack(
         children: [
-          Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              SizedBox(
-                height: 184,
-                width: 162,
-                child: ClipRRect(
-                  borderRadius: BorderRadius.circular(8),
-                  child: Image(
-                    fit: BoxFit.fill,
-                    image: NetworkImage(data.image ?? "N/A"),
-                  ),
-                ),
-              ),
-              Row(
-                children: [
-                  SizedBox(
-                    height: 20,
-                    child: ListView.builder(
-                        scrollDirection: Axis.horizontal,
-                        shrinkWrap: true,
-                        itemBuilder: (context, index) {
-                          return SizedBox(
-                            width: 13,
-                            height: 12,
-                            child: Image.asset(MyIcons.activeStarIcon,
-                                fit: BoxFit.scaleDown),
-                          );
-                        },
-                        itemCount: numberStarActive),
-                  ),
-                  SizedBox(
-                    height: 20,
-                    child: ListView.builder(
-                        scrollDirection: Axis.horizontal,
-                        shrinkWrap: true,
-                        itemBuilder: (context, index) {
-                          return SizedBox(
-                            width: 13,
-                            height: 12,
-                            child: Image.asset(MyIcons.starIcon,
-                                fit: BoxFit.scaleDown),
-                          );
-                        },
-                        itemCount: numberStarNoActive),
-                  ),
-                  const SizedBox(width: 3),
-                  Text(
-                    "${data.star}",
-                    style: const TextStyle(
-                        fontSize: 10,
-                        color: MyColors.colorGray,
-                        fontWeight: FontWeight.normal),
-                  )
-                ],
-              ),
-              Text(
-                data.name,
-                style: const TextStyle(
-                    fontSize: 11,
-                    color: MyColors.colorGray,
-                    fontWeight: FontWeight.normal),
-              ),
-              Text(
-                data.type,
-                maxLines: 1,
-                style: const TextStyle(
-                    fontSize: 16,
-                    overflow: TextOverflow.clip,
-                    color: MyColors.colorBlack,
-                    height: 1,
-                    fontWeight: FontWeight.w600),
-              ),
-              Row(
-                children: [
-                  RichText(
-                    text: TextSpan(
-                      text: 'Color: ',
-                      style: const TextStyle(
-                          fontSize: 11,
-                          height: 1,
-                          fontWeight: FontWeight.normal,
-                          color: MyColors.colorGray),
-                      children: <TextSpan>[
-                        TextSpan(
-                            text: '${data.color}',
-                            style: const TextStyle(color: MyColors.colorBlack)),
-                      ],
-                    ),
-                  ),
-                  const SizedBox(
-                    width: 30,
-                  ),
-                  RichText(
-                    text: TextSpan(
-                      text: 'Size: ',
-                      style: const TextStyle(
-                          fontSize: 11,
-                          height: 1,
-                          fontWeight: FontWeight.normal,
-                          color: MyColors.colorGray),
-                      children: <TextSpan>[
-                        TextSpan(
-                            text: '${data.size}',
-                            style: const TextStyle(color: MyColors.colorBlack)),
-                      ],
-                    ),
-                  )
-                ],
-              ),
-              data.discount == 0.0
-                  ? Text(
-                      "${XUtils.formatPrice(data.originalPrice)}\$ ",
-                      style: const TextStyle(
-                          fontSize: 14,
-                          height: 1.42,
-                          color: MyColors.colorBlack,
-                          fontWeight: FontWeight.w600),
-                    )
-                  : RichText(
-                      text: TextSpan(
-                        text: "${XUtils.formatPrice(data.originalPrice)}\$ ",
-                        style: const TextStyle(
-                            fontSize: 14,
-                            height: 1.42,
-                            decoration: TextDecoration.lineThrough,
-                            color: MyColors.colorGray,
-                            fontWeight: FontWeight.w600),
-                        children: <TextSpan>[
-                          TextSpan(
-                              text:
-                                  "${XUtils.formatPrice(data.currentPrice ?? -1)}\$",
-                              style: const TextStyle(
-                                color: MyColors.colorSaleHot,
-                                decoration: TextDecoration.none,
-                              )),
-                        ],
-                      ),
-                    ),
-            ],
-          ),
+          _belowCard(),
           Container(
             color: MyColors.colorWhite.withOpacity(0.5),
             height: 280,
@@ -393,7 +252,7 @@ class XProductCardFavoriteVertical extends StatelessWidget {
                                   color: MyColors.colorGray),
                               onTap: () => context
                                   .read<FavoriteBloc>()
-                                  .removeProduct(data),
+                                  .removeProductToFavorite(data),
                             )),
                       )
                     ],
