+++ lib/src/modules/home/widgets/new_product.dart
@@ -1,7 +1,11 @@
 import 'package:e_commerce/src/config/themes/my_colors.dart';
+import 'package:e_commerce/src/models/handle.dart';
 import 'package:e_commerce/src/models/products_model.dart';
-import 'package:e_commerce/src/modules/home/logic/product_bloc.dart';
+import 'package:e_commerce/src/models/result.dart';
+import 'package:e_commerce/src/modules/product/logic/product_bloc.dart';
 import 'package:e_commerce/src/widgets/card/product_card_new.dart';
+import 'package:e_commerce/src/widgets/state/state_error_widget.dart';
+import 'package:e_commerce/src/widgets/state/state_loading_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -10,10 +14,11 @@ class NewProduct extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<ProductBloc, ProductState>(
-      builder: (context, state) {
-        final data = state.items ?? [];
-        List<XProduct> items = data.where((e) => e.newProduct == true).toList();
+    return BlocBuilder<ProductBloc, ProductState>(builder: (context, state) {
+      final data = state.items.data ?? [];
+      List<XProduct> items = data.where((e) => e.newProduct == true).toList();
+      XHandle handle = XHandle.result(XResult.success(items));
+      if (handle.isCompleted) {
         return SizedBox(
           height: 337,
           child: Column(
@@ -40,8 +45,12 @@ class NewProduct extends StatelessWidget {
             ],
           ),
         );
-      },
-    );
+      } else if (handle.isLoading) {
+        return const XStateLoadingWidget();
+      } else {
+        return const XStateErrorWidget();
+      }
+    });
   }
 
   Widget _header() {
