+++ lib/src/modules/home/widgets/sale_product.dart
@@ -1,8 +1,11 @@
-
 import 'package:e_commerce/src/config/themes/my_colors.dart';
+import 'package:e_commerce/src/models/handle.dart';
 import 'package:e_commerce/src/models/products_model.dart';
-import 'package:e_commerce/src/modules/home/logic/product_bloc.dart';
+import 'package:e_commerce/src/models/result.dart';
+import 'package:e_commerce/src/modules/product/logic/product_bloc.dart';
 import 'package:e_commerce/src/widgets/card/product_card_sale.dart';
+import 'package:e_commerce/src/widgets/state/state_error_widget.dart';
+import 'package:e_commerce/src/widgets/state/state_loading_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -12,35 +15,42 @@ class SaleProduct extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return BlocBuilder<ProductBloc, ProductState>(builder: (context, state) {
-      final data = state.items ?? [];
+      final data = state.items.data ?? [];
       List<XProduct> items = data.where((e) => e.discount != 0).toList();
-      return SizedBox(
-        height: 337,
-        width: double.infinity,
-        child: Column(
-          children: [
-            _header(),
-            const SizedBox(
-              height: 22,
-            ),
-            Expanded(
-              child: ListView.builder(
-                  scrollDirection: Axis.horizontal,
-                  itemBuilder: (context, index) {
-                    return Padding(
-                      padding: const EdgeInsets.symmetric(horizontal: 10),
-                      child: XProductCardSale(
-                        data: items[index],
-                      ),
-                    );
-                  },
-                  shrinkWrap: true,
-                  physics: const BouncingScrollPhysics(),
-                  itemCount: items.length),
-            ),
-          ],
-        ),
-      );
+      XHandle handle = XHandle.result(XResult.success(items));
+      if (handle.isCompleted) {
+        return SizedBox(
+          height: 337,
+          width: double.infinity,
+          child: Column(
+            children: [
+              _header(),
+              const SizedBox(
+                height: 22,
+              ),
+              Expanded(
+                child: ListView.builder(
+                    scrollDirection: Axis.horizontal,
+                    itemBuilder: (context, index) {
+                      return Padding(
+                        padding: const EdgeInsets.symmetric(horizontal: 10),
+                        child: XProductCardSale(
+                          data: items[index],
+                        ),
+                      );
+                    },
+                    shrinkWrap: true,
+                    physics: const BouncingScrollPhysics(),
+                    itemCount: items.length),
+              ),
+            ],
+          ),
+        );
+      } else if (handle.isLoading) {
+        return const XStateLoadingWidget();
+      } else {
+        return const XStateErrorWidget();
+      }
     });
   }
