+++ lib/src/modules/product/logic/product_state.dart
@@ -1,0 +1,45 @@
+part of 'product_bloc.dart';
+
+class ProductState extends Equatable {
+  final XHandle<List<XProduct>> items;
+  final XHandle<List<XProduct>> searchList;
+  final String searchText;
+
+  final ViewType viewType;
+  final SortBy sortBy;
+  final SizeType sizeType;
+
+  const ProductState(
+      {required this.items,
+      this.sortBy = SortBy.lowToHigh,
+      this.viewType = ViewType.listView,
+      this.sizeType = SizeType.xs,
+      required this.searchList,
+      this.searchText = ''});
+
+  @override
+  List<Object?> get props => [
+        items,
+        searchList,
+        searchText,
+        sortBy,
+        sizeType,
+        viewType,
+      ];
+  ProductState copyWithItem(
+      {XHandle<List<XProduct>>? items,
+      XHandle<List<XProduct>>? searchList,
+      SortBy? sortBy,
+      ViewType? viewType,
+      SizeType? sizeType,
+      String? searchText}) {
+    return ProductState(
+      items: items ?? this.items,
+      searchText: searchText ?? this.searchText,
+      searchList: searchList ?? this.searchList,
+      sortBy: sortBy ?? this.sortBy,
+      viewType: viewType ?? this.viewType,
+      sizeType: sizeType ?? this.sizeType,
+    );
+  }
+}
