+++ lib/src/modules/product_details/widgets/related_products.dart
@@ -1,6 +1,10 @@
 import 'package:e_commerce/src/config/themes/my_colors.dart';
-import 'package:e_commerce/src/modules/home/logic/product_bloc.dart';
-import 'package:e_commerce/src/widgets/card/product_card_gird.dart';
+import 'package:e_commerce/src/models/handle.dart';
+import 'package:e_commerce/src/models/result.dart';
+import 'package:e_commerce/src/modules/product/logic/product_bloc.dart';
+import 'package:e_commerce/src/widgets/card/product_card_vertical.dart';
+import 'package:e_commerce/src/widgets/state/state_error_widget.dart';
+import 'package:e_commerce/src/widgets/state/state_loading_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -10,37 +14,44 @@ class RelatedProducts extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return BlocBuilder<ProductBloc, ProductState>(builder: (context, state) {
-      var data = state.items ?? [];
-      data.shuffle();
-      return SizedBox(
-        height: 294,
-        child: Column(
-          children: [
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 16),
-              child: _header(),
-            ),
-            Expanded(
-              child: Padding(
-                padding: const EdgeInsets.only(left: 16, top: 10),
-                child: ListView.builder(
-                    scrollDirection: Axis.horizontal,
-                    itemBuilder: (context, index) {
-                      return Padding(
-                        padding: const EdgeInsets.symmetric(horizontal: 10),
-                        child: XProductCardGrid(
-                          data: data[index],
-                        ),
-                      );
-                    },
-                    shrinkWrap: true,
-                    physics: const BouncingScrollPhysics(),
-                    itemCount: 5),
+      var items = state.items.data ?? [];
+      items.shuffle();
+      XHandle handle = XHandle.result(XResult.success(items));
+      if (handle.isCompleted) {
+        return SizedBox(
+          height: 294,
+          child: Column(
+            children: [
+              Padding(
+                padding: const EdgeInsets.symmetric(horizontal: 16),
+                child: _header(),
               ),
-            ),
-          ],
-        ),
-      );
+              Expanded(
+                child: Padding(
+                  padding: const EdgeInsets.only(left: 16, top: 10),
+                  child: ListView.builder(
+                      scrollDirection: Axis.horizontal,
+                      itemBuilder: (context, index) {
+                        return Padding(
+                          padding: const EdgeInsets.symmetric(horizontal: 10),
+                          child: XProductCardVertical(
+                            data: items[index],
+                          ),
+                        );
+                      },
+                      shrinkWrap: true,
+                      physics: const BouncingScrollPhysics(),
+                      itemCount: 5),
+                ),
+              ),
+            ],
+          ),
+        );
+      } else if (handle.isLoading) {
+        return const XStateLoadingWidget();
+      } else {
+        return const XStateErrorWidget();
+      }
     });
   }
